requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
tldextract==5.1.2
tenacity==9.0.0
openai==1.43.0
httpx==0.27.2
gspread==6.1.2
google-auth==2.33.0
apify-client==1.8.1
apify-shared==1.1.2
streamlit
pandas==2.2.2
numpy==1.26.4
